cli: config: move helper functions to commands.config module
There are no external callers, so let's make them private.
